unbound: fix problematic migration of interfaces
Interfaces may not be available or disabled.  We can still
give a hint with the tokenizer/type-ahead but we have to
accept the old configuration or else the migration will
always fail as long as bad values are stored.  CSV field
will allow these arbitrary values and deletion thereof if
needed even if we do not follow a strict UX pattern here.

Also remove the Loopback from the selection.  It is always
added automatically to prevent the user breaking local
resolving.