interfaces: improve UX and simplify bridge validation
To be honest the page was horrible.  Originally I wanted to reduce
the bridge validation code using interface_parent_devices() but
ended up renaming a lot of variables from "port" to "dev(ice)" and
started renaming the table headers for clarity which ended up in
splitting the add interface section from the list/modify section
which ended up restyling the whole table for better mobile use and
clear separation of functionality.  Also added a note to the add
section when no devices are available for assignment.

This should in theory prevent a number of support questions and
mistakes seen over the years WRT using this page.