mvc: set Required=Y for GroupNameField
The validation message is never seen anymore and can't be overridden.
While the field made sure to emit a message on empty the result is
that if we ever have to use this group field as optional we would have
to pivot to this approach anyway.

(cherry picked from commit 465f7fa8bb80a526c36d037b0ed675e825550a7c)