wireguard: add dependency for now #6827
Not pretty.  Need to hook wireguard-kmod on life support for
the initial 24.1 at least because we don't know at build time
if the kernel to be installed will have its own kernel module
or not.  Having both works, but the kernel one breaks wireguard-go.

Also implementations for kernel and port kmod seem to be "slightly"
different.