interfaces: DAD sleep on $reload; closes #6913
This is called through rc.linkup exhibiting the issue.
Sidestep the complexity of the situation by fixing the
issue first making it testable and easy to ship in a
stable relese.

For anyone not liking this net.inet6.ip6.dad_count can
be set to "0" to disable the sleep behaviour.  This
needs to be extended one way or another.  More soon.