interfaces: fix the intricate cleanup of link-local VIPs
The VIPs are bound to the scope so we need to add it to the
address lookup.  From the controller emit the same "@" usage
which is used to indicatge a scope, but not no a network
device.  If encountered translate the scope back to a network
device and use the proper "%" syntax.

(cherry picked from commit 16d0a071ad9daa8e8aeec1e31fec8162f3824ff0)