firmware: implement "always reboot" for #6940
The way this works is a little funky...

1. For console calling update script it will ALWAYS reboot,
   but at least it says so.  We simply do not know how much
   will be updated / is available but that is by design
   leaving this as a capable fallback option that will work
   even if the GUI has issues.  That's also true for no
   updates being installed... it reboots by virtue of
   calling the update script into active duty.

2. The GUI and API will have the reboot flag set and know
   this beforehand, but will prevent calling the script
   when it isn't going to do anything.  This makes the reboot
   behaviour better, but ideally we also need to know if we
   need to abort the reboot based on package install progress
   which we currently do not check.

That being said the feature is fine to use when updating the
box via GUI/API/OPNcentral on a regular basis without messing
with the box too much in between.  It's not tailored for micro-
mangement so it should only be used where the behavior fits
the requirement of a consistent system state after firmware
updates (and not hammering the update server every hour for
hotfixes which will make it reboot also).

I guess we need to make further changes, but this is a good
step nonetheless.