System: Gateways: Configuration - Improve migration and fix minor validation issue causing the validation failing. (https://github.com/opnsense/core/pull/6808)
When there are invalid gateway items, it's likely better to skip the ones that we can not migrate in this case. Executing the validation on every record might be expensive, but likely leads to a more consistent end-result. After migration, make sure to remove the old cruft when we end up with at least one gateway.

Without "AllowDynamic" on the intreface selector, we will miss dynamic interfaces like ipsec vti's.