interfaces: split bring_down into dynamic and static parts #6852
Add new variants interface_suspend() and interface_reset() in order
to test the new variants once for rc.linkup and interfaces.php.
Temporary file read isn't optimal but it avoids tainting callers
with additional arguments that might be wrong...  Later we can
switch individual use from interface_bring_down() to the one or
the other and refactor again if it is no longer called directly.

I'm entirely unsure about the PPP on demand primer, but in general
it would make sense to avoid it when the interface is disabled.
The historic way of doing it is priming it for reconfiguration on
'bring-down' so that it can be back up when the line is required.