interfaces: switch some of this back, detect problematic reset
When the interface is not disabled and reset this may be done because
it has been changed and needs to be reconfigured but this will leave
an old static IP on the interface because the address may have been
changed (looking at you, setaddr.php).  The same is likely true for
port reassingment through interfaces_assign.php.