interfaces: refactor interface_bring_down() into interface_reset() and interface_suspend() #6852 (#7040)
Separate use cases to allow for clean non-removal of static addressing
(including CARP VIPs) for use in rc.linkup et al.  While interface_bring_down()
is more or less only being extended with $suspend the journey of this code
change was long and winding. Therefore remove the old ambiguous function
and introduce the new interface_reset() and interface_suspend() variant to
make more sense of the caller's requirements.

interfaces.php is a special case depending on interface enable state so we
pass the $suspend request down to interface_reset() to avoid an if-else
dance.