MVC: flatten complex array structures for better search functionality
implode() dies when provided with arrays in arrays (on development),
so this is a good opportunity to extend the behavior a bit.

The consequence of this commit is that information that is not displayed
in a bootgrid, but still passed in as part of the dataset is also
searchable. This isn't a bad thing, but controllers still have
the option of constraining the amount of fields that should be
searched, as showcased here in interfacesInfoAction().