firmware: disallow TLS 1.0 and 1.1 on business mirror
On the mirror we control prevent the use of "legacy" TLS versions
to allow for better security.  This could be carried over to other
servers but since we don't controll most community servers this
could have negtive impact on availability depending on their config.

A quick sweep shows that all main functions check.sh, update.sh and
upgrade.sh use the launcher so that works as intended even from the
console.  Another audit is required for other firmware scripts and
situations where e.g. opnsense-update is called manually.  Ideally
all should go to launcher.sh or somehow be able to access the
required environment.