core / trust - add simple re-implementation of certctl tool to speedup the process.
Although "openssl rehash" would be more portable, it doesn't seem to match FreeBSD's file layout.
The script in this commit mimics the certctl tool for the relevant parts, but spending a lot less time while doing so.

If at some point in time the certctl tool starts to behave again, we can simply trash this one and go back to the default option.

proposal for https://github.com/opnsense/core/issues/7063