mvc - change exception handling in runMigrations() to avoid mismatches in attributes being silently ignored.
When a migration tries to address an non existing attribute, a InvalidArgumentException is thrown, but only partially noted in the log.
It's likely safer to abort the migration if steps fail and send the error to the log for further inspection.

(cherry picked from commit 14ea1b96e8923ca0a571e73124930f86ecbd63ae)