VPN: WireGuard: Peer generator - Add "Address" field which auto calculates the next available address in the pool (defined by the instance) for both ipv4 and ipv6.
Extend Firewall/Util to iterate over a cidr range, so we can fetch the first available here.
The address fiels should be the "Allowed IPs" in the peer on this instance end. When using allowed IPs ::/0, 0.0.0.0/0 on both ends, the tunnel should work as well, but may people confuse about the routing part.

When the remote end sends traffic from addresses other than the ones configured on the tunnel, one needs to update the peer manually.