system: as a POC move this security setting to the automatic sysctls #7440
This might not work for all values as we would maybe like the have the
FreeBSD default itself to work its magic.  For security relevant things
bootstrapping the value is useful, but for buffer sizes it might not be.

(cherry picked from commit 24f36bf3323bdb08894a8619ab8e2b22ed557539)