interfaces: move sysctls to system requirements related to #5630
Tie them to the IPv6 allow setting which is used by more things
than just the filter, so also move the IPv6 allow to the general
settings for consistency.

There is not much point to tie these sysctls to their former
condition of something set in SLAAC or DHCPv6 connectivity mode
because "ifdisabled" is used and because prior you couldn't turn
it off without a reboot after having tried the IPv6 connetivity.

To make thus super effective refuse to configure IPv6 on interfaces
in these instances by forcing the individual IPv6 interface modes
to 'none'.