KEA: custom DNS for reservation #7310 - unfinished import implementation.
Problem: stream import is done correctly with all variables.
But in validation step the new variables were thrown out
and I have no clue where to add them as "valid".

Additional this is implemented in Base model so it has to be
proofed to be interoperable with all other model usages or
rewritten to other function call.