Add OpenVPN float option
In the classic, legacy server variant there was an option called
"dynamic_ip" which injected the options `persist-remote-ip` and `float`
into the generated configuration file. In the modern variant
`persist-remote-ip` is already available, but `float` isn't available
anymore.

The OpenVPN `float` option is used to allow client to change IP
addresses during an ongoing connection to allow roaming e.g. in mobile
networks or changes in the client WiFi.