webgui: change locking in the webgui restart loosely related to #7649
Emit the new config, but do not write it to the disk yet.
Make sure we can take the config file lock before proceeding.
If we cannot get the lock the work is already being done so
exit in this case.

While here bail early on empty listeners and restructure and
consolidate the config dir path with the resulting config out
of /var/etc.