Improve CPU temperature measurement
There can be a mis-match between the CPU temperature
reported by the operating system and the CPU temperature
shown in the OPNsense dashboard.

This happens for N100, N200 and possibly other CPU types.

The issue seems to be that the sysctl -a command (which
returns around 14,000 lines of output) causes a short-lived
spike in CPU temperature. This spike causes an elevated
temperatue value to be reported in the dashboard.

Avoiding "sysctl -a" means that the dashboard temperature
matches the OS reported temperature.

Current script:

 # /usr/local/opnsense/scripts/system/temperature.sh
 dev.cpu.0.temperature=46.0C
 .
 .
 .

With this change:

 # ./temperature.sh
 dev.cpu.0.temperature=40.0C
 .
 .
 .

This issue was previously reported here:
 https://github.com/opnsense/core/issues/6911