Reporting / rrd - first step trying to split the rrd code in an attempt to make it more flexible and pluggable
Add RRD package with a simple factory class and a basic construct to define different rrd output types we support.
The plan is to migrate the structure first and then take a look at the data collectors, without the code duplication in the rrd.inc file, it should be easier to read and restructure. The RRD\Types namespace should contain a class per output type (e.g. Traffic equals the layout for one interface)