dashboard: widget error recovery
If widgets fail to load for any reason, the tick routine simply continues
without interruption. This commit makes sure that the visual error
state isn't stuck and switches back to widget output when possible.
The easiest way to reproduce this is to simply disconnect from
the network while having the dashboard open.