system: migrate live mode banner to notification
More work needed but as discussed with @swhite2 we want this here
before we work on the UX part.  These types of notifications need
more screen presence and are non-dismissable.

While here kill an oddity assignment with constant() use.

And adjust comments still refering to "LiveCD" since we have a DVD
image nowadays and the live mode works on all install media.