interfaces: add newwanip_map event and deprecate old newwanip one
The new "mapped" event will push a CSV interface list for a batch
reload where possible.  It also feeds the address family requesting
this event so that the listener can decide if this is relevant or
not.  This heavily relies on the type of application or implementation
of the pluggable code underneath and now gives maximum flexibility
without having to call multiple times possibly causing multiple
spurious restarts of the application underneath (as it might be a
monolithic one).

Backport this right away for maximum compatiblity between core,
plugins and external modifications.  We will switftly move all of
our code to this new event.