System/Logging - handle lines with no timestamp (#7889)
Recent feature addition to apply time constraint fails in cases where
a log line does not contain a timestamp, as in the case of multi-line
log entries. This change just moves on when a line contains no
timestamp, as we will eventually find the line that contains the
start of the log entry (and the timestamp).

fixes #7888

(cherry picked from commit 2a1ccae9591c37e4a4c635451beb5e081a8c49c8)