system: refactor route creation for exec_safe() use
While here ignore all return values consistently.  Most of them are
ignored, some of them already on the add case.  Makes little sense
to report errors on a fraction of operations and adding the same route
will unfortunately yield an error anyway so it's not helpful in the
bulk of cases either.