system: adjust CSS for ChartJS v4
maintainAspectRatio has changed to respect the container height.
This has introduced issues as is described in the issue linked in
this commit. Since some of the chart containers rely on maintainAspectRatio
being set to false, setting the aspect-ratio in CSS there will cause
those charts to ignore this setting, causing infinite scaling. These
containers are excluded from the workaround using
canvas-container-noaspectratio. Since these charts do not change
in height, they are not subject to the same degree of responsiveness.

While here, using a chart with an overflowable widget seems unnecessary
and is causing visual artifacts. Eventually we want to split up the
"expansion" logic to their own widgets or configurable setting anyway.