opnsense-beep: serialize access to /dev/speaker
Otherwise one of concurrent calls to `beep' might fail with error message
"beep: open of /dev/speaker for writing: Device busy".

Locking is done in `opnsense-beep' and not in `beep' to preserve tune (avoid
cacophony).  ${BEEPDIR}/low is used as a lock-file as `low' beep is already a
special file: it's default beep.

This might sound like a trivial change, but it might affect system behavior,
e.g. when `opnsense-beep' is used in some shell script with errexit enabled.