dashboard: block display interferes with responsive layouts
chart.js canvas elements default their display to block, which can
cause sub-pixel rounding errors, in turn causing flickering when
the canvas element is interacted with (reproducible on Firefox).

'display: inline-block' fixes this, and while one could specify this
in css, this would have to be combined with !important, as the style
ordering prefers styles either set hardcoded inline or via JS. This
has the obvious downside that a canvas element cannot be hidden anymore,
as it's display setting would always be forced.

The only solution in terms of ordering is to override the inline style
by setting it in the canvas element directly as this commit is doing.

All canvas elements used by widgets which fit the following criteria
must use this inline style:

- a canvas element is used by chart.js
- said canvas lives in a responsive container
- the chart must be resizable in height, which requires the 'canvas-container'
  parent element.