ipsec/openvpn: add deprecation notices for legacy components, closes https://github.com/opnsense/core/issues/8350
Add a note on top of the legacy pages about the end of maintenance, depending on version, this will be version 26.1 ot 26.4 [BE].
In the meantime components will move to plugins. When maintenance ends, it will still be possible to install the component, but may break unexpectedly.

(cherry picked from commit 48ec3b9850969b216a1f8a6b4a71228bb6c5ed36)