mvc: move "lazy loading" option to base model implementation and force usage on run_migrations.php.
In rare cases it is possible to lock the system during boot while drivers are loaded and tunables try to fetch all information from sysctl.
Since we already implemented a lazy loading pattern on the Alias model, it seems to make sense to push this up the chain and reuse it.

For consistency reasons, we should also push the "lazy" attribute when constructing new ModelRelationField types.