system: safety checks for unserialize()
If unserialize() fails for whatever reason, we must assume
an E_WARNING can be issued, false is returned (a literal false
as a valid result is possible, but not relevant in our case) or
a Throwable is thrown.

PHP8.4 will start throwing TypeErrors and ValueErrors here as well,
which are all caught by Throwable.

Since the returned data is never directly used in our code but
intended for enriching purposes, we should have no problems returning
an empty array so the caller fallback is used instead.