Improve dummynet_stats.py parsing of mask descriptor lines.
When using the *-ip6 mask specifiers introduced in #8581, dnctl produces output that dummynet_stats.py fails to correctly parse.

The line in particular is:

```
        mask: proto: 0x00, flow_id: 0x00000000,  ::/0x0000 -> ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/0x0000
```

The line above is currently incorrectly parsed as an IPv6 flow, rather than being ignored. This PR improves the flow parsing by explicitly rejecting lines that describes a mask.