system: recovery should know about wwwony as well
The trouble is the package script can fail or not be called
so we end up with a no-op for chown which ends up with root
rights to dirs while the GUI tries to use wwwonly which does
not even exist worst case.

Also sync with FreeBSD 14.2 except for the /bin/csh -> /bin/sh
change to root user.  We should do this as well, but not for
now.