automation/filter: Do not reload the grid during reconfigureAct
Since we call functions to repopulate the selectpickers during the reconfigureAct
they triggered the changed events to reload the bootgrid and change the url hash.
A new variable was introduced that tracks the reconfigureActInProgress and prevents
spurious reloads during it. After the selectpickers have been repopulated with their
preserved values, a promise is returned which sets the reconfigureActInProgress to
false, unlocking the normal change event behavior during selecpicker usage.