bootgrid: provide more styling possibilities from formatters
While the onRendered callback was already there, it is now wrapped
with the cell component as a parameter for usage in the callback.
This allows a construct such as:

onRendered((cell) => cell.getRow().getElement().style.backgroundColor = 'red');

For example, the value of a cell may be used to determine that
a row should be more/less opaque, such as if an item is disabled.