mvc: ModelRelationField - used cached data when referring the same model, for https://github.com/opnsense/core/issues/9032
There is a small downside in using cached data, as you can't update the same model and use selectors refering the same information without persisting it first.
In practice that's not likely a pattern used in our code base, but using the persisted json file will most likely be quicker than recursively parsing the model to collect the requested data.