firmware: store update and upgrade logs in edge cases; closes #9029
Since /tmp is not persistent and we do not want to keep every single
log for every audit type add a KEEP_LOG parameter to output_done()
and output_reboot() and use it only when we reboot or when the path
ends up in a detectable error.

We are not interested in storing every update log that went ok and
the system keeps running.  LOCKFILE can still be read in that case
unless it was overwritten by user action, in which case the persisent
file would likely also be overwritten.

(cherry picked from commit db65c7b6f37193f4d2ef15e9b2eabfc1e7b88610)