mvc: slightly adjust two getOption() calls in constraints
The xmllint indent wants to turn <check></check> into <check/>
which seems just as spurious so use getOption() $default argument
to pass the expected empty string when not found and change one
other spot to omit the default that is already set by the function
declaration.

While here we can switch a string cast for isEqual().  Unit and
manual tests are fine.