Allow source NAT to use interface networks
The old interface only lists interface addesses, not nets, and the rule parser parser only looks for interface names suffixed by "ip". In pf, interface address is used dynamically by being surrounded by parenthesis and includes the ":0" modifier to exclude aliases. By combining this with the ":network" modifier, source nat can also use an interface's network, dynamically, i.e. "($iface:network:0)".

This is different to one-to-one nat because the source and target networks do not need to be the same size. Any of pf's pool options are valid in this situation.

Signed-off-by: Steve Kerrison <steve.kerrison@jcu.edu.au>