firewall: live view: optimize viewbuffer rendering
Using addData + remove will costs quite some rendering time. Doing a
clear() instead, and then immediately rendering the current state
of the viewbuffer seems to be a lot faster. This also makes sure
that we aren't inserting rows beyond the bufferSize limit, which
serve no purpose, as these have to be removed anyway.