firewall: live log: prevent re-resolving in-flight requests and move host lookup to current filtered view
Previously, it tried to resolve the entire buffer, but this isn't
necessary just to speed up filter times. These can instead be looked
up on the fly, and let the viewmodel column formatter collect
the IPs it sees to be resolved.

This commit also marks in-flight lookups so they don't get re-resolved.

firewall: live log: clean up per-row lookup logic