firewall: live log: only execute redraw on visibility state transition
The IntersectObserver was doing unnecessary work when the general
layout of the table changed, which happens every second in the firewall
live log. Instead, only trigger it when we know our previous event
was the opposite. While here, also remove the default hostname spinner,
as it was causing issues in the detailed rule modal.