interfaces: reduce use of get_real_interface() and $realifXXX variables
Building on the last commit we can get rid of most of the trickery in cases
where we know the device is going to be stored under 'if'.

Although 6rd and 6to4 magic isn't completely gone there may be a future
without it and it's going to be pretty nice.