Interfaces: Neighbors: Automatic Discovery - change list_hosts.py to only query ndp when explicitly asked.
Due to the short timeouts of the ndp entries and the cost of querying them, it's probably best to only support ndp when explicitly asked.
In most cases where we need host data, we currently only support ipv4 anyway, which means a switch to the new entrypoint (list_hosts.py script) should be relatively harmless.