VPN: IPsec: Security Policy Database - extend ModelRelationField so it can optionally disable caching. closes https://github.com/opnsense/core/issues/9365
In some cases caching can be problematic, for example when rows are dynamically generated or when the cache is persisted while the model hasn't reached it final state yet.
The latter seems to be the case here. When referring connection, we actually mean the description of the related field, which may not be persisted in the model yet as an earlier field triggered the data to be stored (locals.local.connection for example).

This commit adds "DisableCache" as ModelRelationField option and uses it in SPDs.SPD.connection_child

(cherry picked from commit cbd6ea95a759e61bb9a8475b877ebdabf59141e6)