Firewall: Aliases - use new hostdiscovery (with arp/ndp fallback) in mac type aliases.
While here, cleanup some redundant code, if a mac address is in the local cache, the local cache should be complete at anytime.
Technically, for legacy ndp, this might be a bit worse than before, but as hostdiscovery is more complete, that should be a small price to pay.

Eventually, when hostdiscoverty is the standard, we should be able to ditch the /tmp/alias_filter_arp.cache construction as hostdiscovery has its own database.