firewall: remove mute from 1 execution call #9325
Mute is used when we know the operation could fail but these don't
look like it and shouldn't.  In the shaper case the script even ends
using exit 0 without a set -e so that's going to be fine all the time.

pfctl -d|-e errors when already in the right mode, which is a bit
unelegant.