mvc: rewrite the old Shell class according to our current standards #9325
Remove instantiation and add exec_safe() as well as shell_safe(), run_safe()
and pass_safe() on top of those.  Replace the use in the Monit migration.

This is based on my work carried out over the years in util.inc regarding
exec_safe() and mwexecf().  The next commit will go as far as replacing the
functions there with the class implementations to avoid a drift between
backend and mvc code.