interfaces: interface_carp_configure() conversion
The use of str_replace+addslashes+escapeshellarg is very odd:

https://github.com/pfsense/pfsense/commit/942fdd555964d48
https://redmine.pfsense.org/issues/213

To avoid bugs leave this for now.  escapeshellarg should be able
to handle this on its own.  The password parsing is not more special
than anything else passed to the command line unless there is a bug
somewhere else which I doubt.