mvc: introduce defaultConfig property for AppConfig
Related to testing there is a side effect from snatching the
default config.xml if no other configuration was found.  We
want the tests to be independent of runtime so this is a first
step towards it.  While here replace other uses of the default
file acquire and almost completely fix the permission issue
on default bootstrap.