Interfaces: Neighbors: Automatic Discovery - by default, only use non-promiscuous mode, final point for https://github.com/opnsense/core/pull/9354
Added to the roadmap as well, as it seemed to be missing.

This new service is a prerequisite for the new (and faster) mac type aliases (https://github.com/opnsense/core/commit/b2a30fc5606ce2d6c781ae9b7282b83e8ec35ac3) and is a requirement for IPv6 captive portal in a future release.